added line-pattern to file reader
note: this is only used import a single field (the whole line).  the pattnern is used to match the line.  if it is a match, it's read into transformalize, otherwise it is appended to the previous line.  this feature may change.  it could either skip anything doesn't match the pattern, or do this append to previous line thing it's doing now.